Move 'levelized_random_access' class fully to its own file
